+++ NEWS.md
@@ -1,3 +1,15 @@
+# dfadjust 1.1.0
+
+## New Features
+
+- Allow argument `ell` to be shorter than covariate dimension. In this case,
+  `ell` specifies which subset of covariates to compute standard errors for.
+
+## Minor improvements and fixes
+
+- Use `collapse::fsum` instead of `tapply` calls to improve speed
+- Check that covariates are not collinear, drop the collinear ones
+
 # dfadjust 1.0.5
 
 ## Minor improvements and fixes